Irs system us

+++ Modules/EmployeeSalaryItems/Database/Migrations/2023_05_29_111257_create_deduction_details_table.php
@@ -19,8 +19,8 @@ return new class extends Migration
                 ->foreignId('employee_salary_item_id')
                 ->constrained('employee_salary_items', 'id')
                 ->cascadeOnDelete();
-            $table->integer('employee_amount')->nullable();
-            $table->integer('government_or_company_amount')->nullable();
+            $table->float('employee_amount')->nullable();
+            $table->float('government_or_company_amount')->nullable();
             $table->integer('start_percentage')->nullable();
             $table->integer('end_percentage')->nullable();
             $table->timestamps();
